Small implementation of a host
A very small script that pings all the servers in the config.yaml (located in the root of the server, in this scenario), asks for the rooms, and returns a response from each server in a manner that's compatible with SRB2. I'm not sure how rooms with conflicting numbers will be handled, hence issue #3, so I'm hesitating to make any of my own implementation of such without any second opinion. A note is that the output of this PHP file in a browser returns unformatted text (no line breaks)