Rework Kart Netgame model
In order to migrate to a more flexible database backend, the SRB2Kart's
Netgame Model has been reworked to utilize the newly-added
*PDOSingleton*. The latter has been added to bundle possible connections
between multiple APIs and can be seen as a generic database backend.

Relying on PDO over ODBC also appears to be more flexible and less
error-prone, according to the PHP documentation.